feat(sdk): add client.notebooks covering every platform-side notebook command
list/get/status/plan/create/wait/start/stop/delete/events/follow_events/
lifecycle/metrics/realtime_metrics/quotas/save_image/estimate_image_size/
cancel_save_image/wait_image_ready. Notebook create inputs, status
vocabulary, output projections and the workload quota query core move
to inspire/services with public names; CLI and SDK build identical
create payloads (verified against `notebook create --dry-run`).

+++ cli/inspire/platform/web/browser_api/images.py
@@ -318,8 +318,8 @@ _IMAGE_FAILED_STATES = {
 def wait_for_image_ready(
     image_id: str,
     session: Optional[WebSession] = None,
-    timeout: int = 600,
-    poll_interval: int = 5,
+    timeout: float = 600,
+    poll_interval: float = 5,
 ) -> CustomImageInfo:
     """Wait for a custom image to reach a terminal success state.
